gene_e/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohappyeyeballs=2.3.6=pypi_0
aiohttp=3.10.3=pypi_0
annotated-types=0.7.0=pypi_0
anthropic=0.34.0=pypi_0
anyio=4.4.0=pypi_0
black=24.8.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.7.2=haa95532_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
discord-py=2.4.0=pypi_0
distro=1.9.0=pypi_0
filelock=3.15.4=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.6.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.24.5=pypi_0
idna=3.7=pypi_0
jiter=0.5.0=pypi_0
libffi=3.4.4=hd77b12b_1
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
nodeenv=1.9.1=pypi_0
openssl=3.0.14=h827c3e9_0
packaging=24.1=pypi_0
pathspec=0.12.1=pypi_0
pip=24.2=py311haa95532_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pyright=1.1.376=pypi_0
python=3.11.9=he1021f5_0
pyyaml=6.0.2=pypi_0
setuptools=72.1.0=py311haa95532_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
tk=8.6.14=h0416ee5_0
tokenizers=0.20.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=1.26.19=pypi_0
vc=14.40=h2eaa2aa_0
vs2015_runtime=14.40.33807=h98bb1dd_0
wheel=0.43.0=py311haa95532_0
xz=5.4.6=h8cc25b3_1
zlib=1.2.13=h8cc25b3_1
